refactor(server): split id and token helpers
